Expanded CreatureInfo [Complete]
Additional detail has been added to all creatures. The complete list of attributes now stored can be found in dma\dnd\creature_info.py. The new attributes are:
-treasure types
-speed (for ground, water, air, climbing, burrowing, in webs)
-armor class
-number and damage ranges of basic attacks
-magic resistance
-alignment
-size class
-origin sourcebook
-creature hierarchies (creature subtypes know thier parent, and parents know their subtypes)
-alternate names for creatures

Unit tests for CreatureInfo definitions were added

In the process I also identified a number of additional attributes that may be useful to store in a future CreatureInfo pass. These can be found in creature_tbd_fields.txt.

Next up will be setting up the django CreatureInfo models to store the new attributes in the database, and then updating the creature detail template to display the new information.